feat: Add ViewPager and adapter. Add some panel edit controls.
Still working on adding all of the panel edit controls, getting these
to work with the view pager and adapter. Found a solution so that
the view of image currently displayed to user is the view in mImageView.
Problem here was that the adapter loads multiple views, and the right
one was not always (mostly never) the one view found in mImageView.